+++ src/src/app/books/page.tsx
@@ -1,7 +1,7 @@
 import { BookList } from '../ui/books/BookList';
 import { Title }  from '../ui/utils/Title';
 
-const BooksForm = () => {
+const BooksPage = () => {
   return (
     <div>
       <Title value="本一覧"/>
@@ -10,4 +10,4 @@ const BooksForm = () => {
   );
 }
 
-export default BooksForm;
+export default BooksPage;
